requirements.txt
# --- CORE ML & AUDIO ---
torch>=2.0.0
torchvision
torchaudio
numpy<2.1.0
scipy
librosa
soundfile

# --- MODELLI ---
munch
pyyaml
nltk
jieba
cn2an
pypinyin
eng-to-ipa
phonemizer
unidecode

# --- ANALISI & VISUALIZZAZIONE ---
scikit-learn
matplotlib
seaborn
tqdm

# --- UTILS ---
requests
python-dotenv
onnxruntime